requirements.txt
fastapi
uvicorn
openai
requests
python-dotenv
redis>=5.0
